Develop

+++ app/Http/Controllers/EntriesController.php
@@ -22,6 +22,7 @@ use Auth;
 use Validator;
 use Input;
 use Redirect;
+// use Permission;
 use Helper;
 use Log;
 use App\Models\Entry;
@@ -155,7 +156,9 @@ class EntriesController extends Controller
     public function getCreate(Request $request)
     {
         //log::debug("getCreate: entered >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>");
-
+        if(!\Permission::checkPermission('add-entry-permission', $request->whitelabel_group)) {
+            return view('errors.403');       
+        }
         $post_types = ['want' => 'I want', 'have' => 'I have'];
 
         $request->session()->put('upload_key', str_random(15));
@@ -354,7 +357,11 @@ class EntriesController extends Controller
      */
     public function getEdit(Request $request, $entryID)
     {
+            
 
+        if(!\Permission::checkPermission('edit-any-entry-permission', $request->whitelabel_group)) {
+            return view('errors.403');       
+        }
         //log::debug("getEdit: entered >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>");
 
         // This should be pulled into a helper or macro
@@ -362,11 +369,10 @@ class EntriesController extends Controller
 
         if ($entry = Entry::find($entryID)) {
             $user = Auth::user();
-
-            if ($request->user()->cannot('update-entry', $entry)) {
+            if ($request->user()->cannot('update-entry', [$entry, $request->whitelabel_group])) {
+                
                 return redirect()->route('home')->with('error', trans('general.entries.messages.not_allowed'));
             }
-
             $image = \DB::table('media')
                 ->where('entry_id', '=', $entryID)
                 ->first();
@@ -488,12 +494,13 @@ class EntriesController extends Controller
      */
     public function postEdit(Request $request, $entryID)
     {
+        
         //log::debug("postEdit: entered >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>");
 
         if ($entry = Entry::find($entryID)) {
             $user = Auth::user();
 
-            if ($request->user()->cannot('update-entry', $entry)) {
+            if ($request->user()->cannot('update-entry', [$entry, $request->whitelabel_group])) {
                 abort(403);
             }
 
@@ -583,7 +590,7 @@ class EntriesController extends Controller
      * @param $entryID
      * @return String JSON
      */
-    public function postAjaxDelete($entryID)
+    public function postAjaxDelete(Request $request,$entryID)
     {
         try {
             $entry = $this->dispatchNow(new DeleteEntry($entryID));
@@ -610,8 +617,11 @@ class EntriesController extends Controller
      * @param $entryID
      * @return Redirect
      */
-    public function postDelete($entryID)
+    public function postDelete(Request $request, $entryID)
     {
+        if(!\Permission::checkPermission('delete-any-entry-permission', $request->whitelabel_group)) {
+            return view('errors.401');       
+        }
         try {
             $this->dispatchNow(new DeleteEntry($entryID));
         } catch (ModelNotFoundException $exception) {
@@ -706,19 +716,19 @@ class EntriesController extends Controller
             $entries->textSearch(e(Input::get('search')));
         }
 
-        if (Input::has('offset')) {
-            $offset = e(Input::get('offset'));
-        }
-        else {
-            $offset = 0;
-        }
+        // if (Input::has('offset')) {
+        //     $offset = e(Input::get('offset'));
+        // }
+        // else {
+        //     $offset = 0;
+        // }
 
-        //if (Input::has('limit')) { pagination stuff and we may need it one day - dsl
+        // if (Input::has('limit')) { 
         //    $limit = e(Input::get('limit'));
-        //}
-        //else {
-        //    $limit = 50;
-        //}
+        // }
+        // else {
+        //    $limit = 10;
+        // }
 
         $allowed_columns = [
             'title',
@@ -734,9 +744,12 @@ class EntriesController extends Controller
         $sort = in_array(Input::get('sort'), $allowed_columns) ? Input::get('sort') : 'entries.created_at';
         $order = Input::get('order') == 'asc' ? 'asc' : 'desc';
 
+        // $entries = $entries->orderBy($sort, $order)->paginate($offset, $limit)->get();
+
         $entries = $entries->orderBy($sort, $order)->get();
-        $count = $entries->count();
-        // $entries = $entries->skip($offset)->take($limit)->get(); pagination stuff and we may need it one day - dsl
+        // $count = $entries->count(); 
+        $count = count($entries);        
+               
 
         $rows = [];
 
@@ -908,6 +921,7 @@ class EntriesController extends Controller
             'rows'     => $rows,
             'pois'     => $allPois,
             'viewType' => $entryLayout,
+            // 'entries'  => $entries->total(),
         ];
     }
